MovieDatabase/requirements.txt
psycopg2==2.6.2
django==1.10.5
djangorestframework==3.5.4